requirements_v4.txt
# Additional requirements for DATA_BOT v4
# Extending v3 requirements with enterprise features

# GraphQL support
graphene==3.3
strawberry-graphql==0.212.0
graphql-core==3.2.3

# OpenSearch support
opensearch-py==2.4.0

# Machine Learning & Clustering
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
transformers==4.35.0
sentence-transformers==2.2.2
umap-learn==0.5.4
hdbscan==0.8.33

# Enhanced Admin Interface
streamlit==1.28.0
plotly==5.17.0
dash==2.14.1

# Kubernetes support
kubernetes==28.1.0
pyyaml==6.0.1

# Enhanced async support
asyncpg==0.29.0
redis==5.0.1

# Monitoring and metrics
prometheus-client==0.19.0
structlog==23.2.0

# Security enhancements
cryptography==41.0.7
passlib==1.7.4
python-jose==3.3.0

# Additional utilities
rich==13.7.0
typer==0.9.0
watchdog==3.0.0